+++ packages/maestro/test/maestro_test.dart
@@ -1,8 +1,22 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:maestro/maestro.dart';
 
-class DefaultPerfomer with Performer {}
+class DefaultPerfomer implements Performer {
+  @override
+  void attach(Score score) {}
+
+  @override
+  void detach() {}
+
+  @override
+  FutureOr<void> play() {}
+
+  @override
+  void remix(DefaultPerfomer old) {}
+}
 
 void main() {
   group('Maestro', () {
@@ -72,8 +86,7 @@ void main() {
       expect(value, equals(168));
     });
 
-    testWidgets('rebuilt if initial value changed and if it is not a performer',
-        (tester) async {
+    testWidgets('rebuilt if value changed', (tester) async {
       int buildCount = 0;
 
       final Widget child = Builder(
@@ -89,25 +102,6 @@ void main() {
 
       await tester.pumpWidget(Maestro(84, child: child));
       expect(buildCount, equals(2));
-    });
-
-    testWidgets('not rebuilt if initial value changed and if it is a performer',
-        (tester) async {
-      int buildCount = 0;
-
-      final Widget child = Builder(
-        builder: (context) {
-          buildCount++;
-          Maestro.listen<DefaultPerfomer>(context);
-          return const SizedBox();
-        },
-      );
-
-      await tester.pumpWidget(Maestro(DefaultPerfomer(), child: child));
-      expect(buildCount, equals(1));
-
-      await tester.pumpWidget(Maestro(DefaultPerfomer(), child: child));
-      expect(buildCount, equals(1));
     });
   });
 }