+++ packages/maestro/test/performer_test.dart
@@ -1,31 +1,47 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:maestro/maestro.dart';
 
 class _DelegatedPerformer implements Performer {
-  const _DelegatedPerformer(this.onAttached, this.onDispose);
+  const _DelegatedPerformer(
+    this.onAttach,
+    this.onPlay,
+    this.onRemix,
+    this.onDetach,
+  );
 
-  final void Function(Score score) onAttached;
-  final VoidCallback onDispose;
+  final void Function(Score score) onAttach;
+  final void Function() onPlay;
+  final void Function(_DelegatedPerformer old) onRemix;
+  final VoidCallback onDetach;
 
   @override
-  void attach(Score score) => onAttached(score);
+  void attach(Score score) => onAttach(score);
 
   @override
-  void detach() => onDispose();
+  FutureOr<void> play() async => onPlay();
+
+  @override
+  void remix(_DelegatedPerformer old) => onRemix(old);
+
+  @override
+  void detach() => onDetach();
 }
 
 void main() {
   group('Performer', () {
     testWidgets('methods are called when they should', (tester) async {
-      final List<bool> attached = [false, false];
-      final List<bool> disposed = [false, false];
+      final List<String> logs = <String>[];
 
       final List<_DelegatedPerformer> performers = List.generate(
         2,
         (i) => _DelegatedPerformer(
-          (score) => attached[i] = true,
-          () => disposed[i] = true,
+          (score) => logs.add('attach $i'),
+          () => logs.add('play $i'),
+          (_DelegatedPerformer old) => logs.add('remix $i'),
+          () => logs.add('detach $i'),
         ),
       );
 
@@ -39,24 +55,28 @@ void main() {
         ),
       );
 
-      expect(attached[0], equals(true));
-      expect(disposed[0], equals(false));
+      expect(logs, <String>[
+        'attach 0',
+        'play 0',
+      ]);
+      logs.clear();
 
       await tester.pumpWidget(
         Maestros(
           [
             const Maestro(1),
-            Maestro(performers[1], key: UniqueKey()),
+            Maestro(performers[1]),
           ],
           child: const SizedBox(),
         ),
       );
 
-      expect(attached[0], equals(true));
-      expect(disposed[0], equals(true));
-
-      expect(attached[1], equals(true));
-      expect(disposed[1], equals(false));
+      expect(logs, <String>[
+        'attach 1',
+        'remix 1',
+        'detach 0',
+      ]);
+      logs.clear();
 
       await tester.pumpWidget(
         const Maestros(
@@ -67,11 +87,9 @@ void main() {
         ),
       );
 
-      expect(attached[0], equals(true));
-      expect(disposed[0], equals(true));
-
-      expect(attached[1], equals(true));
-      expect(disposed[1], equals(true));
+      expect(logs, <String>[
+        'detach 1',
+      ]);
     });
   });
 }